Allow users to input their Twitter app keys
Previous version required user to input their keys into the program prior to running it.  Now, a user, who is not a programmer, can input their keys into the UI.  
Also, changed "library" commands to "require" so that if the app is run locally, the program will install necessary packages automatically if they are not already on the user's machine.